+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.0
+
+- Update webpack 5.
+- Add more field return (https://github.com/kwent/react-google-contacts/pull/2). Thanks @nicks78.
+- maxResults props added (https://github.com/kwent/react-google-contacts/pull/8). Thanks @moshfiqrony.
+
 ## 0.0.1
 
 - First alpha release.